price_calculation.jsファイルの作成
