+++ app/controllers/items_controller.rb
@@ -1,7 +1,6 @@
 class ItemsController < ApplicationController
   before_action :authenticate_user!, only: [:new, :create]
   def index
-    @items = Item.all
   end
 
   def new
@@ -13,7 +12,6 @@ class ItemsController < ApplicationController
     if @item.save
       redirect_to items_path(@item)
     else
-      Rails.logger.info @item.errors.full_messages # エラー内容をターミナルに出力
       render :new
     end
   end
