+++ app/models/item.rb
@@ -7,7 +7,6 @@ class Item < ApplicationRecord
   belongs_to_active_hash :shipping_date_estimate
 
   belongs_to :user
-  has_one :purchase
   has_one_attached :image
 
   validates :image, presence: true
